Fixing bug where the 0th account would always get logged in
The login form would select account i where i was the index of the tapped username. But if a smaller list of usernames was displayed (because of filtering), the displayed list and the backing list would no longer have the same indices. This now uses the actual text to compare so it should be slower (technically) but more correct.